+++ src/web/tests/tests.py
@@ -13,16 +13,13 @@ from .factories import (
     UserProfileFactory,
     ServiceFactory,
 )
-from .models import Massage, Customer, User, Service
-from .importer import (
+from ..models import Massage, Customer, User, Service
+from ..importer import (
     therapist_import,
     services_import,
     customer_import,
     massage_import,
     update_or_create_w_logging,
-    single_massage_import,
-    massage_appointments,
-    massage_date_comparison_with_wp_db,
 )
 
 tz = pytz.timezone("Europe/Ljubljana")
@@ -1302,645 +1299,6 @@ class ImportDataTest(TestCase):
         self.assertEqual(
             massage.start.astimezone(tz).isoformat(), expected_date.isoformat()
         )
-
-    def test_single_massage_import(self):
-        therapist = UserProfileFactory()
-        massage1 = MassageFactory.build(
-            external_id=30, start=datetime.datetime(2023, 5, 5, 17, 0, 0)
-        )
-        massage2 = MassageFactory.build(
-            external_id=270, start=datetime.datetime(2023, 4, 7, 9, 0, 0)
-        )
-        customer1 = CustomerFactory.build()
-        customer2 = CustomerFactory.build()
-        service1 = ServiceFactory()
-        service2 = ServiceFactory()
-
-        data1 = {
-            "message": "Successfully retrieved appointments",
-            "data": {
-                "appointments": {
-                    "2023-06-07": {
-                        "date": "2023-06-07",
-                        "appointments": [
-                            {
-                                "id": 576,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer1.external_id,
-                                            "firstName": customer1.name,
-                                            "lastName": customer1.surname,
-                                            "email": customer1.email,
-                                            "phone": customer1.phone,
-                                        },
-                                        "status": massage1.status,
-                                        "price": service1.price,
-                                        "appointmentId": massage1.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage1.status,
-                                "serviceId": service1.external_id,
-                                "providerId": therapist.external_id,
-                                "bookingStart": massage1.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage1.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                    "2023-05-05": {
-                        "date": "2023-05-05",
-                        "appointments": [
-                            {
-                                "id": 555,
-                                "bookings": [
-                                    {
-                                        "id": 270,
-                                        "customerId": 41,
-                                        "customer": {
-                                            "id": customer2.external_id,
-                                            "firstName": customer2.name,
-                                            "lastName": customer2.surname,
-                                            "birthday": None,
-                                            "email": customer2.email,
-                                            "phone": customer2.phone,
-                                        },
-                                        "status": massage2.status,
-                                        "price": service2.price,
-                                        "appointmentId": massage2.external_id,
-                                        "persons": 1,
-                                        "duration": 5400,
-                                        "created": "2023-04-4 09:20:20",
-                                    }
-                                ],
-                                "status": massage2.status,
-                                "serviceId": service2.external_id,
-                                "providerId": therapist.external_id,
-                                "bookingStart": massage2.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage2.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                },
-            },
-        }
-
-        # changed massage date of massage2 from 2023-04-7 at 9:00 to 2023-10-7 at 10:00
-        data2 = {
-            "message": "Successfully retrieved appointment",
-            "data": {
-                "appointment": {
-                    "id": massage2.external_id,
-                    "bookings": [
-                        {
-                            "id": 691,
-                            "customerId": customer2.external_id,
-                            "customer": None,
-                            "status": massage2.status,
-                            "price": service2.price,
-                            "appointmentId": massage2.external_id,
-                            "persons": 1,
-                            "duration": 5400,
-                            "created": "2023-04-4 09:20:20",
-                        }
-                    ],
-                    "status": massage2.status,
-                    "serviceId": service2.external_id,
-                    "providerId": therapist.external_id,
-                    "bookingStart": "2023-10-7 10:00:00",
-                    "bookingEnd": massage2.end.strftime("%Y-%m-%d %H:%M:%S"),
-                    "type": "appointment",
-                    "isRescheduled": None,
-                    "isFull": None,
-                    "resources": [],
-                },
-                "recurring": [],
-            },
-        }
-
-        UserProfileFactory(external_id="42")
-
-        massage_count = Massage.objects.all().count()
-
-        customer_import(data1)
-        massage_import(data1)
-
-        single_massage_import(data2)
-
-        # check that date and hour of massage1 has not changed
-        massage = Massage.objects.get(external_id=30)
-        tz = pytz.timezone("Europe/Berlin")
-        expected_date = datetime.datetime(2023, 5, 5, 17, 0, 0)
-        expected_date = tz.localize(expected_date)
-        self.assertEqual(
-            massage.start.astimezone(tz).isoformat(), expected_date.isoformat()
-        )
-
-        massage = Massage.objects.get(external_id=270)
-        # check if massage is only changed and not added
-        self.assertEqual(Massage.objects.all().count(), massage_count + 2)
-        # check if start date  of massage2 is changed from 2023-04-7 at 9:00 to 2023-10-7 at 10:00
-        tz = pytz.timezone("Europe/Berlin")
-        expected_date = datetime.datetime(2023, 10, 7, 10, 0)
-        expected_date = tz.localize(expected_date)
-        self.assertEqual(
-            massage.start.astimezone(tz).isoformat(), expected_date.isoformat()
-        )
-
-        # when start date is changed to future date, does it register in db
-        massage = Massage.objects.latest("id")
-        massage.refresh_from_db()
-        self.assertEqual(massage.start, expected_date)
-
-    def test_get_massage_appointments(self):
-        """
-        Get a list of external_id from the wodrpess API call
-        """
-        therapist1 = UserProfileFactory(user__first_name="Jane").user
-        self.client.force_login(therapist1)
-
-        customer1 = CustomerFactory(name="Brian", main_concern="car accident")
-        customer2 = CustomerFactory(name="Alice", main_concern="bike accident")
-        customer3 = CustomerFactory(name="David", main_concern="fall off tree")
-
-        massage1 = MassageFactory(
-            therapist=therapist1,
-            customer=customer1,
-            start=datetime.datetime(2023, 4, 6, 16, 0, 0).astimezone(tz=tz),
-            status="approved",
-            external_id=5,
-        )
-        massage2 = MassageFactory(
-            therapist=therapist1,
-            customer=customer2,
-            start=datetime.datetime(2023, 4, 6, 18, 0, 0).astimezone(tz=tz),
-            status="canceled",
-            external_id=7,
-        )
-        massage3 = MassageFactory(
-            therapist=therapist1,
-            customer=customer3,
-            start=datetime.datetime(2023, 4, 7, 17, 0, 0).astimezone(tz=tz),
-            status="approved",
-            external_id=8,
-        )
-
-        data = {
-            "message": "Successfully retrieved appointments",
-            "data": {
-                "appointments": {
-                    "2023-04-06": {
-                        "date": "2023-04-06",
-                        "appointments": [
-                            {
-                                "id": 576,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer1.external_id,
-                                            "firstName": customer1.name,
-                                            "lastName": customer1.surname,
-                                            "email": customer1.email,
-                                            "phone": customer1.phone,
-                                        },
-                                        "status": massage1.status,
-                                        "price": massage1.service.price,
-                                        "appointmentId": massage1.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage1.status,
-                                "serviceId": massage1.service.external_id,
-                                "providerId": massage1.therapist.userprofile.external_id,
-                                "bookingStart": massage1.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage1.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                            {
-                                "id": 577,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer2.external_id,
-                                            "firstName": customer2.name,
-                                            "lastName": customer2.surname,
-                                            "email": customer2.email,
-                                            "phone": customer2.phone,
-                                        },
-                                        "status": massage2.status,
-                                        "price": massage2.service.price,
-                                        "appointmentId": massage2.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage2.status,
-                                "serviceId": massage2.service.external_id,
-                                "providerId": massage2.therapist.userprofile.external_id,
-                                "bookingStart": massage2.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage2.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                    "2023-04-07": {
-                        "date": "2023-04-07",
-                        "appointments": [
-                            {
-                                "id": 579,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer3.external_id,
-                                            "firstName": customer3.name,
-                                            "lastName": customer3.surname,
-                                            "email": customer3.email,
-                                            "phone": customer3.phone,
-                                        },
-                                        "status": massage3.status,
-                                        "price": massage3.service.price,
-                                        "appointmentId": massage3.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage3.status,
-                                "serviceId": massage3.service.external_id,
-                                "providerId": massage3.therapist.userprofile.external_id,
-                                "bookingStart": massage3.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage3.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                },
-            },
-        }
-
-        massage_appointments_in_wp_db = massage_appointments(data)
-        self.assertEqual(massage_appointments_in_wp_db, [5, 7, 8])
-
-    @freeze_time("2023-04-06 13:21:34", tz_offset=2)
-    def test_compare_massage_date_with_wp_db(self):
-        therapist1 = UserProfileFactory(user__first_name="Jane").user
-        self.client.force_login(therapist1)
-
-        customer1 = CustomerFactory()
-        customer2 = CustomerFactory()
-        customer3 = CustomerFactory()
-        customer4 = CustomerFactory()
-        customer5 = CustomerFactory()
-
-        massage1 = MassageFactory(
-            therapist=therapist1,
-            customer=customer1,
-            start=datetime.datetime(2023, 4, 6, 16, 0, 0).astimezone(tz=tz),
-            status="approved",
-            external_id=5,
-        )
-        massage2 = MassageFactory(
-            therapist=therapist1,
-            customer=customer2,
-            start=datetime.datetime(2023, 4, 6, 18, 0, 0).astimezone(tz=tz),
-            status="canceled",
-            external_id=7,
-        )
-        massage3 = MassageFactory(
-            therapist=therapist1,
-            customer=customer3,
-            start=datetime.datetime(2023, 4, 7, 17, 0, 0).astimezone(tz=tz),
-            status="approved",
-            external_id=8,
-        )
-        massage4 = MassageFactory(
-            therapist=therapist1,
-            customer=customer4,
-            start=datetime.datetime(2023, 4, 7, 20, 0, 0).astimezone(tz=tz),
-            status="approved",
-            external_id=10,
-        )
-        massage5 = MassageFactory(
-            therapist=therapist1,
-            customer=customer5,
-            start=datetime.datetime(2023, 5, 17, 21, 0, 0).astimezone(tz=tz),
-            status="not canceled",
-            external_id=15,
-        )
-        # does not contain massage4(ex_id10)
-        # contains massage5(ex_id15)
-        data1 = {
-            "message": "Successfully retrieved appointments",
-            "data": {
-                "appointments": {
-                    "2023-04-06": {
-                        "date": "2023-04-06",
-                        "appointments": [
-                            {
-                                "id": 576,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer1.external_id,
-                                            "firstName": customer1.name,
-                                            "lastName": customer1.surname,
-                                            "email": customer1.email,
-                                            "phone": customer1.phone,
-                                        },
-                                        "status": massage1.status,
-                                        "price": massage1.service.price,
-                                        "appointmentId": massage1.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage1.status,
-                                "serviceId": massage1.service.external_id,
-                                "providerId": massage1.therapist.userprofile.external_id,
-                                "bookingStart": massage1.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage1.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                            {
-                                "id": 577,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer2.external_id,
-                                            "firstName": customer2.name,
-                                            "lastName": customer2.surname,
-                                            "email": customer2.email,
-                                            "phone": customer2.phone,
-                                        },
-                                        "status": massage2.status,
-                                        "price": massage2.service.price,
-                                        "appointmentId": massage2.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage2.status,
-                                "serviceId": massage2.service.external_id,
-                                "providerId": massage2.therapist.userprofile.external_id,
-                                "bookingStart": massage2.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage2.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                    "2023-04-07": {
-                        "date": "2023-04-07",
-                        "appointments": [
-                            {
-                                "id": 579,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer3.external_id,
-                                            "firstName": customer3.name,
-                                            "lastName": customer3.surname,
-                                            "email": customer3.email,
-                                            "phone": customer3.phone,
-                                        },
-                                        "status": massage3.status,
-                                        "price": massage3.service.price,
-                                        "appointmentId": massage3.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage3.status,
-                                "serviceId": massage3.service.external_id,
-                                "providerId": massage3.therapist.userprofile.external_id,
-                                "bookingStart": massage3.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage3.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                            {
-                                "id": 579,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer5.external_id,
-                                            "firstName": customer5.name,
-                                            "lastName": customer5.surname,
-                                            "email": customer5.email,
-                                            "phone": customer5.phone,
-                                        },
-                                        "status": massage5.status,
-                                        "price": massage5.service.price,
-                                        "appointmentId": massage5.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage5.status,
-                                "serviceId": massage5.service.external_id,
-                                "providerId": massage5.therapist.userprofile.external_id,
-                                "bookingStart": massage5.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage5.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                },
-            },
-        }
-
-        # added massage4
-        # does not contain massage1 and massage5
-        wordpress_data = {
-            "message": "Successfully retrieved appointments",
-            "data": {
-                "appointments": {
-                    "2023-04-06": {
-                        "date": "2023-04-06",
-                        "appointments": [
-                            {
-                                "id": 577,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer2.external_id,
-                                            "firstName": customer2.name,
-                                            "lastName": customer2.surname,
-                                            "email": customer2.email,
-                                            "phone": customer2.phone,
-                                        },
-                                        "status": massage2.status,
-                                        "price": massage2.service.price,
-                                        "appointmentId": massage2.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage2.status,
-                                "serviceId": massage2.service.external_id,
-                                "providerId": massage2.therapist.userprofile.external_id,
-                                "bookingStart": massage2.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage2.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                    "2023-04-07": {
-                        "date": "2023-04-07",
-                        "appointments": [
-                            {
-                                "id": 579,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer3.external_id,
-                                            "firstName": customer3.name,
-                                            "lastName": customer3.surname,
-                                            "email": customer3.email,
-                                            "phone": customer3.phone,
-                                        },
-                                        "status": massage3.status,
-                                        "price": massage3.service.price,
-                                        "appointmentId": massage3.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage3.status,
-                                "serviceId": massage3.service.external_id,
-                                "providerId": massage3.therapist.userprofile.external_id,
-                                "bookingStart": massage3.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage3.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                            {
-                                "id": 600,
-                                "bookings": [
-                                    {
-                                        "id": 345,
-                                        "customerId": 333,
-                                        "customer": {
-                                            "id": customer4.external_id,
-                                            "firstName": customer4.name,
-                                            "lastName": customer4.surname,
-                                            "email": customer4.email,
-                                            "phone": customer4.phone,
-                                        },
-                                        "status": massage4.status,
-                                        "price": massage4.service.price,
-                                        "appointmentId": massage4.external_id,
-                                        "persons": 1,
-                                        "duration": 3600,
-                                        "created": "2023-03-31 15:15:50",
-                                    }
-                                ],
-                                "status": massage4.status,
-                                "serviceId": massage4.service.external_id,
-                                "providerId": massage4.therapist.userprofile.external_id,
-                                "bookingStart": massage4.start.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                                "bookingEnd": massage4.end.strftime(
-                                    "%Y-%m-%d %H:%M:%S"
-                                ),
-                            },
-                        ],
-                    },
-                },
-            },
-        }
-
-        # import data1 to local db
-        massage_import(data1)
-        massage5 = Massage.objects.get(external_id=15)
-        self.assertEqual(massage5.status, "not_canceled")
-
-        # import wordpress_data,
-        massage_import(wordpress_data)
-        massage4 = Massage.objects.get(external_id=10)
-        self.assertEqual(massage4.status, "approved")
-        self.assertEqual(massage4.external_id, 10)
-
-        # massage = Massage.objects.latest("id")
-        # massage.refresh_from_db()
-        # self.assertEqual(massage.start, expected_date
-
-        # # check wp IDs against local IDs
-        wordpress_external_id = massage_appointments(wordpress_data)
-
-        # only_local_db should be [5] only_in_wp_db should be [4]
-        only_in_local_db, only_in_wordpress_db = massage_date_comparison_with_wp_db(
-            wordpress_external_id
-        )
-        # self.assertEqual(massage.external_id, 15)
-        # print(only_in_local_db)
-        # print(only_in_wordpress_db)
-
-        # self.assertEqual((only_in_local_db, only_in_wordpress_db), ([5, 7, 8], [10]))
 
 
 class LogDataTest(TestCase):
